Updated footer component
Reach us still redirects to github URL
